+++ src/interfaces/libpq/fe-misc.c
@@ -128,7 +128,8 @@ pqPutBytes(const char *s, size_t nbytes, PGconn *conn)
 		{
 			printfPQExpBuffer(&conn->errorMessage,
 							  libpq_gettext("could not flush enough data (space available: %d, space needed %d)\n"),
-					  Max(conn->outBufSize - conn->outCount, 0), nbytes);
+							  (int) Max(conn->outBufSize - conn->outCount, 0),
+							  (int) nbytes);
 			return EOF;
 		}
 		/* fixup avail for while loop */
